fixed the issue with form text on both user/admin
Also fixed the image component diaplay function. Now it keeps on fetching untill the image is displayed, and also added a placeholder in case if the resource fails to reload. since we are using multer so our server it self is handling the file transfers which can sometime behave un expectedly depends on network situations.